+++ test/aampcli/AampcliGet.cpp
@@ -358,7 +358,7 @@ void Get::ShowHelpGet(){
 char * Get::getCommandRecommender(const char *text, int state)
 {
     char *name;
-    static int len;
+    static size_t len;
     static std::vector<std::string>::iterator itr;
 
     if (!state)
